sched: Force sleep on consecutive sched_yields
If a task sched_yields to itself continuously, force the task to
sleep in sched_yield. This will lower the CPU load of this task
thereby lowering the cpu frequency and improving power.

Added a stat variable to track how many times we sleep due these
consecutive sched_yields. Also added sysctl knobs to control the number
of consecutive sched_yields before which the sleep kicks in and the
duration fo the sleep in us.

Bug 1424617

Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-by: Wen Yi <wyi@nvidia.com>
Reviewed-by: Peter Zu <pzu@nvidia.com>
Reviewed-by: Diwakar Tundlam <dtundlam@nvidia.com>
[neo: Include cleanups by @Dorimanx . Thanks.]
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>