acpuclock-7627: fixing table selection in select_freq_plan()
In 'select_freq_plan()' there's a special check to select
the correct table for 1008 Mhz CPU's, because of an conflict
between MSM-7627AA and MSM-8625 boards.

However this check is wrong:

	if (pll_mhz[ACPU_PLL_2] == 245)

...should be:

	if (pll_mhz[ACPU_PLL_1] == 245)

We don't want to check PLL_2 but PLL_1 being '245' or not.

Due of this bug the wrong frequency table is getting selected
for 1 GHz MSM-8625 CPUs with "PLL_2=245".

Previously-detected-by: Peter Jung <peter.jung@lge.com>
Signed-off-by: moddingg33k <moddingg33k@t-online.de>