arm: arch_timer: set memory mapped timer interrupt as IRQF_TIMER
The memory mapped timer is used as a broadcast timer to wake the core for timer interrupts when the arch timer might not be functional. When interrupt is not marked as IRQF_NO_SUSPEND, the interrupt gets disabled during the suspend_device_irqs() callback in the suspend path. If a core were to enter a idle low power mode which relies on broadcast timer to process the interrupt, the core is never woken up for timer interrupts. Mark the interrupt with IRQF_TIMER which marks this interrupt as a timer interrupt and also marks it as IRQF_NO_SUSPEND

Signed-off-by: moddingg33k <moddingg33k@t-online.de>